working on adding and editing food

+++ app/src/main/java/com/davidemodolo/insulyna/food/Food.kt
@@ -1,3 +1,3 @@
 package com.davidemodolo.insulyna.food
 
-data class Food(var name: String, var carbo: Float, var piece: Boolean)
+data class Food(var name: String, var carbo: Int, var piece: Boolean)
